Further refactoring of formats
