+++ opentile/formats/__init__.py
@@ -1,0 +1,19 @@
+#    Copyright 2021 SECTRA AB
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+
+from opentile.formats.ome import OmeTiffTiler
+from opentile.formats.ndpi import NdpiTiler
+from opentile.formats.histech import HistechTiffTiler
+from opentile.formats.philips import PhilipsTiffTiler
+from opentile.formats.svs import SvsTiler
