+++ opentile/formats/ndpi/ndpi_tile.py
