+++ opentile/formats/svs/svs_metadata.py
