--- opentile/ndpi/__init__.py
@@ -1,1 +1,0 @@
-from opentile.ndpi.ndpi_tiler import NdpiTiler
