--- opentile/philips/__init__.py
@@ -1,1 +1,0 @@
-from opentile.philips.philips_tiff_tiler import PhilipsTiffTiler
