--- opentile/svs/__init__.py
@@ -1,1 +1,0 @@
-from opentile.svs.svs_tiler import SvsTiler
